Refactor to encapsulate incomes logic
