+++ src/blocks/Expenses/ExpenseCard/ExpenseCard.jsx
@@ -1,17 +1,21 @@
-import Icon from "../Icon";
+import Icon from "../../../components/Icon";
 import "./ExpenseCard.css";
 
 const ExpenseCard = (props) => {
-    return <div className="expense">
-        {props.icon ? <div className={`expense__color expense__color--${props.color}`}>
-            <Icon icon={props.icon}/>
-        </div> : null}
-        <div className="expense__left">
-            <p className="expense__text--bold">{props.title}</p>
-            <p className="expense__text--regular expense__date">{props.date}</p>
+  return (
+    <div className="expense">
+      {props.icon ? (
+        <div className={`expense__color expense__color--${props.color}`}>
+          <Icon icon={props.icon} />
         </div>
-        <p className="expense__text--bold">{props.amount}</p>
-    </div>;
-}
+      ) : null}
+      <div className="expense__left">
+        <p className="expense__text--bold">{props.title}</p>
+        <p className="expense__text--regular expense__date">{props.date}</p>
+      </div>
+      <p className="expense__text--bold">{props.amount}</p>
+    </div>
+  );
+};
 
 export default ExpenseCard;
