+++ src/blocks/Incomes/IncomeCard/index.js
@@ -1,0 +1,3 @@
+import IncomeCard from "./IncomeCard";
+
+export default IncomeCard;
