+++ src/blocks/Incomes/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Incomes";
